separated the logic of the navigation stack from the NavigationManager
