+++ src/Wpf.Ui/Services/Internal/NavigationManager.cs
@@ -1,6 +1,5 @@
 ﻿#nullable enable
 using System;
-using System.Buffers;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
@@ -15,12 +14,11 @@ namespace Wpf.Ui.Services.Internal;
 
 internal sealed class NavigationManager : IDisposable
 {
+    private readonly NavigationStackManager _navigationStackManager;
     private readonly Frame _frame;
     private readonly INavigationItem[] _navigationItems;
     private readonly FrameworkElement?[] _instances;
     private readonly IPageService? _pageService;
-    private readonly ArrayPool<INavigationItem> _arrayPool = ArrayPool<INavigationItem>.Create();
-    private readonly List<INavigationItem> _navigationStackHistory = new();
 
     private bool _isBackwardsNavigated;
     private bool _addToNavigationStack;
@@ -28,10 +26,11 @@ internal sealed class NavigationManager : IDisposable
 
     public bool CanGoBack => History.Count > 1;
     public readonly List<int> History = new();
-    public readonly ObservableCollection<INavigationItem> NavigationStack = new();
+    public ObservableCollection<INavigationItem> NavigationStack => _navigationStackManager.NavigationStack;
 
     public NavigationManager(Frame frame, IPageService? pageService, INavigationItem[] navigationItems)
     {
+        _navigationStackManager = new NavigationStackManager();
         _instances = new FrameworkElement[navigationItems.Length];
 
         _navigationItems = navigationItems;
@@ -43,8 +42,7 @@ internal sealed class NavigationManager : IDisposable
 
     public void Dispose()
     {
-        NavigationStack.Clear();
-        _navigationStackHistory.Clear();
+        _navigationStackManager.Dispose();
         ClearCache();
     }
 
@@ -125,72 +123,16 @@ internal sealed class NavigationManager : IDisposable
             return false;
         }
 
-        switch (NavigationStack.Count)
-        {
-            case > 0 when NavigationStack[NavigationStack.Count -1] == item:
-                return false;
-            case 0:
-                NavigationStack.Add(item);
-                break;
-        }
+        if (!_navigationStackManager.AddFirstItemAndCheckIfNavigatingToCurrentItem(item))
+            return false;
 
-        AddToNavigationStack(item);
+        _navigationStackManager.AddToNavigationStack(item, _addToNavigationStack, _isBackwardsNavigated);
         ActivateItem(item);
         AddToHistory(itemId);
 
         _frame.Navigate(instance);
-        return true;
-    }
-
-    private void AddToNavigationStack(INavigationItem item)
-    {
-        if (_isBackwardsNavigated && item.WasInBreadcrumb)
-        {
-            if (_navigationStackHistory.Count > 1)
-            {
-                for (var i = 0; i < _navigationStackHistory.Count - 1; i++)
-                {
-                    _addToNavigationStack = true;
-
-                    var historyItem = _navigationStackHistory[i];
-                    historyItem.WasInBreadcrumb = false;
-                    AddToNavigationStack(historyItem);
-                }
-
-                _navigationStackHistory.Clear();
-            }
-
-            item.WasInBreadcrumb = false;
-            _addToNavigationStack = true;
-        }
-
-        if (_addToNavigationStack && !NavigationStack.Contains(item))
-        {
-            item.WasInBreadcrumb = true;
-            NavigationStack.Add(item);
-        }
-
-        if (!item.IsHidden && !_addToNavigationStack)
-        {
-            NavigationStack[0].IsActive = false;
-            NavigationStack[0] = item;
-
-            ClearNavigationStack(1);
-        }
-
-        var navigationStackCount = NavigationStack.Count;
-        if (navigationStackCount > 1)
-        {
-            var navItem = NavigationStack[NavigationStack.Count - 2];
-            if (navItem.IsHidden)
-                navItem.IsActive = false;
-
-            var index = NavigationStack.IndexOf(item);
-            if (index < navigationStackCount - 1 && _navigationStackHistory.Count == 0)
-                ClearNavigationStack(++index);
-        }
-
         _addToNavigationStack = false;
+        return true;
     }
 
     private void ActivateItem(INavigationItem item)
@@ -278,31 +220,6 @@ internal sealed class NavigationManager : IDisposable
 
         _instances[itemData.itemId] = element;
         return _instances[itemData.itemId]!;
-    }
-
-    private void ClearNavigationStack(int navigationStackItemIndex)
-    {
-        var navigationStackCount = NavigationStack.Count;
-        var length = navigationStackCount - navigationStackItemIndex;
-        var buffer = _arrayPool.Rent(length);
-
-        int i = 0;
-        for (int j = navigationStackItemIndex; j <= navigationStackCount - 1; j++)
-        {
-            buffer[i] = NavigationStack[j];
-            i++;
-        }
-
-        for (var index = 0; index < length; index++)
-        {
-            var item = buffer[index];
-            NavigationStack.Remove(item);
-
-            if (length > 1)
-                _navigationStackHistory.Add(item);
-        }
-
-        _arrayPool.Return(buffer, true);
     }
 
     private int GetItemId(Func<INavigationItem, bool> prediction)
